Updated version of the last project
Here, the explanatory variable used in the model is different from last project. In this updated version, the explanatory that tried to be used are whether to play at home or away, who is the opponent, and the expected statistics from each club. The reason why i changed the explanatory variable is that i think, by using the new explanatory variable, the model would be more useful, because the model could predict the outcome before the game is played